backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-dotenv
google-generativeai
playwright